device_name can now be overridden by cli arguments
The device name is now optional in the config, and a `--device-name` CLI
parameter allows specifying the device name that will be used.

This allows using the same configuration file with multiple devices,
provided that the devices are known up front.

Note that you will need to spawn one evremap process per device
to remap multiple devices at the same time.

Thanks to @innovate-invent for making a first pass at this:
closes: https://github.com/wez/evremap/pull/59

refs: https://github.com/wez/evremap/issues/17