Add --wait-for-device flag
This is a simple approach for the case where the device isn't
plugged in on startup.

It works by polling the devices in a loop every few seconds.

It is probably better and more robust to use udev rules for
this purpose, but this approach will do for a quick and dirty
ad-hoc user.

closes: #35
closes: #46
closes: #60